# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "RCPCH Digital Growth API",
        "description": "Returns SDS and centiles for child growth measurements using growth references. Currently provides calculations based on the UK-WHO, Turner's Syndrome and Trisomy-21 references.",
        "version": "5.0.0"
    },
    "paths": {
        "/uk-who/calculation": {
            "post": {
                "tags": [
                    "uk-who"
                ],
                "summary": "Uk Who Calculation",
                "description": "## UK-WHO Centile and SDS Calculations\n\n* These are the 'standard' centiles for children in the UK. It uses a hybrid of the WHO and UK90 datasets.\n* For non-UK use you may need the WHO-only or CDC charts which we do not yet support, but we may add if demand is there.  Please contact us.\n* Returns a single centile/SDS calculation for the selected `measurement_method`.\n* Gestational age correction will be applied automatically if appropriate according to the gestational age at birth data supplied.\n* Available `measurement_method`s are: `height`, `weight`, `bmi`, or `ofc` (OFC = occipitofrontal circumference = 'head circumference').\n* Note that BMI must be precalculated for the `bmi` function.\n* Dates will discard anything after first 'T' in `YYYY-MM-DDTHH:MM:SS.milliseconds+TZ` etc\n* Optional Bone age data associated with a height can be passed:\n*   - `bone_age` as a float in years\n*   - `bone_age_sds` and `bone_age_centile` as floats\n*   - `bone_age_type` as one of `greulich-pyle`, `tanner-whitehouse-ii`, `tanner-whitehouse-iiI`, `fels`, `bonexpert`\n* Optional events can be passed in as a list of strings - each list is associated with a measurement",
                "operationId": "uk_who_calculation_uk_who_calculation_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MeasurementRequest",
                                "examples": [
                                    {
                                        "birth_date": "2020-04-12",
                                        "observation_date": "2028-06-12",
                                        "observation_value": 115,
                                        "sex": "female",
                                        "gestation_weeks": 40,
                                        "gestation_days": 0,
                                        "measurement_method": "height",
                                        "bone_age": 10,
                                        "bone_age_centile": 98,
                                        "bone_age_sds": 2.0,
                                        "bone_age_text": "This bone age is advanced",
                                        "bone_age_type": "greulich-pyle",
                                        "events_text": [
                                            "Growth hormone start",
                                            "Growth Hormone Deficiency diagnosis"
                                        ]
                                    }
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeasurementObject"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/uk-who/bulk-calculation": {
            "post": {
                "tags": [
                    "uk-who"
                ],
                "summary": "Uk Who Bulk Calculation",
                "operationId": "uk_who_bulk_calculation_uk_who_bulk_calculation_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkMeasurementRequest",
                                "examples": [
                                    {
                                        "measurement_method": "height",
                                        "birth_date": "2020-04-12",
                                        "sex": "female",
                                        "observations": [
                                            {
                                                "observation_date": "2028-06-12",
                                                "observation_value": 115
                                            },
                                            {
                                                "observation_date": "2028-12-12",
                                                "observation_value": 130
                                            }
                                        ]
                                    }
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BulkMeasurementObject"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/uk-who/chart-coordinates": {
            "post": {
                "tags": [
                    "uk-who"
                ],
                "summary": "Uk Who Chart Coordinates",
                "description": "## UK-WHO Chart Coordinates data.\n\n* Returns coordinates for constructing the lines of a traditional growth chart, in JSON format\n* Requires a sex ('male' or 'female' lowercase) and a measurement_method ('height', 'weight' ,'bmi', 'ofc')\n* If custom centiles/sds collections (individually or as a collection) are required, accepts a list of float values (up to 15) as centile_format parameter\n* The is_sds boolean flag (default false) specifies if the custom list is of SDS or centiles.\n* In addition to the custom list, \"cole-nine-centiles\" or \"three-percent-centiles\" can be specified which are standard collections.\n* If no centile_format is supplied, \"cole-nine-centiles\" are returned as a default.",
                "operationId": "uk_who_chart_coordinates_uk_who_chart_coordinates_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChartCoordinateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Centile_Data"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/uk-who/fictional-child-data": {
            "post": {
                "tags": [
                    "uk-who"
                ],
                "summary": "Fictional Child Data",
                "description": "## UK-WHO Fictional Child Data Endpoint\n\n* Generates synthetic data for demonstration or testing purposes",
                "operationId": "fictional_child_data_uk_who_fictional_child_data_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FictionalChildRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/MeasurementObject"
                                    },
                                    "type": "array",
                                    "title": "Response Fictional Child Data Uk Who Fictional Child Data Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/who/calculation": {
            "post": {
                "tags": [
                    "who"
                ],
                "summary": "Who Calculation",
                "description": "## WHO Centile and SDS Calculations\n\n* These are the 'standard' centiles for children in North America.\n* Returns a single centile/SDS calculation for the selected `measurement_method`.\n* Gestational age correction will be applied automatically if appropriate according to the gestational age at birth data supplied.\n* Available `measurement_method`s are: `height`, `weight`, `bmi`, or `ofc` (OFC = occipitofrontal circumference = 'head circumference').\n* Note that BMI must be precalculated for the `bmi` function.\n* Dates will discard anything after first 'T' in `YYYY-MM-DDTHH:MM:SS.milliseconds+TZ` etc\n* Optional Bone age data associated with a height can be passed:\n*   - `bone_age` as a float in years\n*   - `bone_age_sds` and `bone_age_centile` as floats\n*   - `bone_age_type` as one of `greulich-pyle`, `tanner-whitehouse-ii`, `tanner-whitehouse-iiI`, `fels`, `bonexpert`\n* Optional events can be passed in as a list of strings - each list is associated with a measurement",
                "operationId": "who_calculation_who_calculation_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MeasurementRequest",
                                "examples": [
                                    {
                                        "birth_date": "2020-04-12",
                                        "observation_date": "2028-06-12",
                                        "observation_value": 115,
                                        "sex": "female",
                                        "gestation_weeks": 40,
                                        "gestation_days": 0,
                                        "measurement_method": "height",
                                        "bone_age": 10,
                                        "bone_age_centile": 98,
                                        "bone_age_sds": 2.0,
                                        "bone_age_text": "This bone age is advanced",
                                        "bone_age_type": "greulich-pyle",
                                        "events_text": [
                                            "Growth hormone start",
                                            "Growth Hormone Deficiency diagnosis"
                                        ]
                                    }
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeasurementObject"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/who/bulk-calculation": {
            "post": {
                "tags": [
                    "who"
                ],
                "summary": "Who Bulk Calculation",
                "operationId": "who_bulk_calculation_who_bulk_calculation_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkMeasurementRequest",
                                "examples": [
                                    {
                                        "measurement_method": "height",
                                        "birth_date": "2020-04-12",
                                        "sex": "female",
                                        "gestation_weeks": 40,
                                        "gestation_days": 0,
                                        "observations": [
                                            {
                                                "observation_date": "2028-06-12",
                                                "observation_value": 115
                                            },
                                            {
                                                "observation_date": "2028-12-12",
                                                "observation_value": 130
                                            }
                                        ]
                                    }
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BulkMeasurementObject"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/who/chart-coordinates": {
            "post": {
                "tags": [
                    "who"
                ],
                "summary": "Who Chart Coordinates",
                "description": "## WHO Chart Coordinates data.\n\n* Returns coordinates for constructing the lines of a traditional growth chart, in JSON format\n* Requires a sex ('male' or 'female' lowercase) and a measurement_method ('height', 'weight' ,'bmi', 'ofc')\n* If custom centiles/sds collections (individually or as a collection) are required, accepts a list of float values (up to 15) as centile_format parameter\n* The is_sds boolean flag (default false) specifies if the custom list is of SDS or centiles.\n* In addition to the custom list, \"cole-nine-centiles\" or \"three-percent-centiles\" can be specified which are standard collections.\n* If no centile_format is supplied, \"cole-nine-centiles\" are returned as a default.",
                "operationId": "who_chart_coordinates_who_chart_coordinates_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChartCoordinateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Centile_Data"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/who/fictional-child-data": {
            "post": {
                "tags": [
                    "who"
                ],
                "summary": "Fictional Child Data",
                "description": "## WHO Fictional Child Data Endpoint\n\n* Generates synthetic data for demonstration or testing purposes",
                "operationId": "fictional_child_data_who_fictional_child_data_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FictionalChildRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/MeasurementObject"
                                    },
                                    "type": "array",
                                    "title": "Response Fictional Child Data Who Fictional Child Data Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MeasurementObject": {
                "properties": {
                    "provenance": {
                        "$ref": "#/components/schemas/Provenance"
                    },
                    "birth_data": {
                        "$ref": "#/components/schemas/BirthData"
                    },
                    "measurement_dates": {
                        "$ref": "#/components/schemas/MeasurementDates"
                    },
                    "child_observation_value": {
                        "$ref": "#/components/schemas/ChildObservationValue"
                    },
                    "measurement_calculated_values": {
                        "$ref": "#/components/schemas/MeasurementCalculatedValues"
                    },
                    "plottable_data": {
                        "$ref": "#/components/schemas/PlottableData"
                    },
                    "bone_age": {
                        "$ref": "#/components/schemas/BoneAge"
                    },
                    "events_data": {
                        "$ref": "#/components/schemas/EventsData"
                    }
                },
                "type": "object",
                "required": [
                    "provenance",
                    "birth_data",
                    "measurement_dates",
                    "child_observation_value",
                    "measurement_calculated_values",
                    "plottable_data",
                    "bone_age",
                    "events_data"
                ],
                "title": "MeasurementObject"
            },
...
        }
    }
}